requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
pandas
pymysql
pymupdf
python-dotenv
PyMuPDF
sentence-transformers  
transformers  
langchain-community
mysql-connector-python
openai
langchain_openai


uvicorn[standard]
pydantic
python-dotenv